// src/generated/models/Pm_documentsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Pm_documentsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Pm_documentsstatecode = keyof typeof Pm_documentsstatecode;
export const Pm_documentsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Pm_documentsstatuscode = keyof typeof Pm_documentsstatuscode;

export interface Pm_documentsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  pm_documentid: string;
  pm_documenttitle: string;
  pm_entityguid?: string;
  pm_module?: string;
  statecode: Pm_documentsstatecode;
  statuscode?: Pm_documentsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Pm_documents extends Pm_documentsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  pm_file?: string;
  pm_file_name?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}
export type Pm_documentsFileColumnName = 'pm_file';

export type Pm_documentsUploadColumnName = Pm_documentsFileColumnName;